Add destination as factor in camera target
The way this works is that champ + mouse + dest are combined via
weighted average to calculate target position of camera. Right now they
all start at 1.0 each (i.e. equal contribution) but the mouse and dest
do fall off the further you are from the champion (this also sort of
fixes the minimap jitter).

Later we can probably adjust the default weights so that we can add more
importance to the dest position or the mouse position, see how each
feels.