More signatures + patching
Signatures of the instructions modifying the cameraX/Y when you click on
the minimap have been added. There are still 3 or 4 of them (:o) I need
to add for eliminating totally the camera movement from the client when
you click the minimap, so we can replace it if we want to.
I don't know if we will use this one day, but if it's the case it will
be there.