2 memory errors fixed + champ hovered detected + view shared
2 memory errors crashed the program at the end, it sometimes unpatched
uncorrectly the client.

The champion hovered is now contained in this->entity_hovered ; it
cannot contain the player champion, only allies or ennemies champions

When you hover a champion, a shared view is created (as if you push Fx)
; I think it is really useful because most of the time, you want to keep
an eye on the ennemy.

You can press space to unshared the camera and restore the default view