Fix #20, "hint" entity added see description
Issue #20 is now fixed

The "Hint entity" is a new weight that modifies slightly the camera
position when you hover an ally/ennemy champion by sharing the current
camera position with the entity's one. The "hint entity" is kept until
the user press Space, F1, or if the entity dies. I need to improve it by
disabling it if the entity stays in the fog for too long

Focus entity is the weight when you *left click* on an ally/ennemy
champion. It is kept until the same conditions than the "hint entity"

So if you hover and left click on a champion, you should be able to
follow precisely this champion